feat: implement custom rrule scalar in sdl, migrated startTeamPrompt to sdl
