+++ packages/server/graphql/rootMutation.ts
@@ -108,7 +108,6 @@ import startCheckIn from './mutations/startCheckIn'
 import startDraggingReflection from './mutations/startDraggingReflection'
 import startRetrospective from './mutations/startRetrospective'
 import startSprintPoker from './mutations/startSprintPoker'
-import startTeamPrompt from './mutations/startTeamPrompt'
 import toggleTeamDrawer from './mutations/toggleTeamDrawer'
 import updateAgendaItem from './mutations/updateAgendaItem'
 import updateAzureDevOpsDimensionField from './mutations/updateAzureDevOpsDimensionField'
@@ -235,7 +234,6 @@ export default new GraphQLObjectType<any, GQLContext>({
       startRetrospective,
       startSprintPoker,
       setTaskHighlight,
-      startTeamPrompt,
       updateAgendaItem,
       updateCommentContent,
       updateCreditCard,
